requirements/dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --all-extras --output-file='requirements\dev-requirements.txt'
#
alabaster==0.7.13
    # via sphinx
babel==2.12.1
    # via sphinx
bokeh==2.4.3
    # via PyICe-adi (pyproject.toml)
cairocffi==1.6.1
    # via cairosvg
cairosvg==2.7.0
    # via PyICe-adi (pyproject.toml)
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via cairocffi
chardet==3.0.4
    # via requests
colorama==0.4.6
    # via
    #   pytest
    #   sphinx
contourpy==1.1.0
    # via matplotlib
coverage[toml]==7.3.0
    # via pytest-cov
cssselect2==0.7.0
    # via cairosvg
cycler==0.11.0
    # via matplotlib
defusedxml==0.7.1
    # via cairosvg
deprecated==1.2.12
    # via PyICe-adi (pyproject.toml)
docutils==0.18.1
    # via
    #   sphinx
    #   sphinx-rtd-theme
et-xmlfile==1.1.0
    # via openpyxl
exceptiongroup==1.1.2
    # via pytest
fonttools==4.42.0
    # via matplotlib
graphviz==0.20.1
    # via PyICe-adi (pyproject.toml)
hightime==0.2.1
    # via
    #   nifgen
    #   nitclk
idna==2.8
    # via requests
imagesize==1.4.1
    # via sphinx
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via
    #   bokeh
    #   sphinx
jsonc-parser==1.1.5
    # via PyICe-adi (pyproject.toml)
kiwisolver==1.4.4
    # via matplotlib
labcomm==0.0.2
    # via PyICe-adi (pyproject.toml)
markupsafe==2.1.3
    # via jinja2
matplotlib==3.7.1
    # via PyICe-adi (pyproject.toml)
minimalmodbus==2.1.1
    # via PyICe-adi (pyproject.toml)
nifgen==1.4.5
    # via PyICe-adi (pyproject.toml)
nitclk==1.4.5
    # via nifgen
numpy==1.21.6
    # via
    #   bokeh
    #   contourpy
    #   matplotlib
    #   pandas
    #   patsy
    #   scikit-rf
    #   scipy
    #   statsmodels
openpyxl==3.1.2
    # via PyICe-adi (pyproject.toml)
packaging==23.1
    # via
    #   bokeh
    #   matplotlib
    #   pytest
    #   sphinx
    #   statsmodels
pandas==1.3.5
    # via
    #   PyICe-adi (pyproject.toml)
    #   scikit-rf
    #   statsmodels
patsy==0.5.3
    # via statsmodels
pillow==9.5.0
    # via
    #   bokeh
    #   cairosvg
    #   matplotlib
pluggy==1.2.0
    # via pytest
pybind11==2.11.1
    # via PyICe-adi (pyproject.toml)
pycairo==1.23.0
    # via PyICe-adi (pyproject.toml)
pycparser==2.21
    # via cffi
pygments==2.15.1
    # via sphinx
pyparsing==3.1.1
    # via matplotlib
pyserial==3.4
    # via
    #   PyICe-adi (pyproject.toml)
    #   minimalmodbus
pyside2==5.15.2.1
    # via PyICe-adi (pyproject.toml)
pysocks==1.7.1
    # via PyICe-adi (pyproject.toml)
pytest==7.4.0
    # via
    #   PyICe-adi (pyproject.toml)
    #   pytest-cov
    #   pytest-mock
pytest-cov==4.1.0
    # via PyICe-adi (pyproject.toml)
pytest-mock==3.11.1
    # via PyICe-adi (pyproject.toml)
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-usbtmc==0.8
    # via PyICe-adi (pyproject.toml)
python-vxi11==0.9
    # via PyICe-adi (pyproject.toml)
pytz==2023.3
    # via pandas
pyusb==1.2.1
    # via PyICe-adi (pyproject.toml)
pyvisa==1.12.0
    # via PyICe-adi (pyproject.toml)
pywin32==306 ; platform_system == "Windows"
    # via PyICe-adi (pyproject.toml)
pyyaml==6.0.1
    # via bokeh
requests==2.22.0
    # via
    #   PyICe-adi (pyproject.toml)
    #   sphinx
scikit-rf==0.26.0
    # via PyICe-adi (pyproject.toml)
scipy==1.7.3
    # via
    #   PyICe-adi (pyproject.toml)
    #   scikit-rf
    #   statsmodels
shiboken2==5.15.2.1
    # via pyside2
six==1.16.0
    # via
    #   patsy
    #   python-dateutil
snowballstemmer==2.2.0
    # via sphinx
sphinx==5.3.0
    # via
    #   PyICe-adi (pyproject.toml)
    #   sphinx-rtd-theme
    #   sphinxcontrib-jquery
sphinx-rtd-theme==1.2.2
    # via PyICe-adi (pyproject.toml)
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jquery==4.1
    # via sphinx-rtd-theme
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
statsmodels==0.13.2
    # via PyICe-adi (pyproject.toml)
tinycss2==1.2.1
    # via
    #   cairosvg
    #   cssselect2
tomli==2.0.1
    # via
    #   coverage
    #   pytest
tornado==6.2
    # via bokeh
typing-extensions==4.7.1
    # via
    #   bokeh
    #   pyvisa
urllib3==1.25.11
    # via requests
webencodings==0.5.1
    # via
    #   cssselect2
    #   tinycss2
windows-curses==2.3.1 ; platform_system == "Windows"
    # via PyICe-adi (pyproject.toml)
wrapt==1.15.0
    # via deprecated
xlrd==2.0.1
    # via PyICe-adi (pyproject.toml)
xlsxwriter==1.2.9
    # via PyICe-adi (pyproject.toml)
